UK bridleway designation not highway=bridleway
Although ism has a concept of highway=bridleway in the uk this has no
correspondence to quality of path it is just a right of way.  So only
use it to set access rights not as the basis of speed mapping.  This
should allow bikes to use routes that have already got an appropriate
highway mapping from descriptive term.  This means we probably need to
look closer at “No Physical Manifestation”