UK footpath not highway=footway, No Physical Manifestation => higway=nohighway
Although osm has a concept of highway=footway in the uk this has no
correspondence to quality of path it is just a right of way. So only
use it to set access rights not as the basis of speed mapping.  Also
map No Physical Manifestation as a highway type so it is accessible to
mountain bikes and other bikes even though this is not a concept
understood by osm.